# gen/openapiv2/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "apptemplate/v1/models.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AppItemsService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/appitems": {
      "get": {
        "summary": "ListAppItems returns all available appitems",
        "operationId": "AppItemsService_ListAppItems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListAppItemsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new appitem",
        "operationId": "AppItemsService_CreateAppItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAppItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAppItemRequest"
            }
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      }
    },
    "/v1/appitems/{appitem.id}": {
      "patch": {
        "summary": "GetAppItem returns a specific appitem with metadata",
        "operationId": "AppItemsService_UpdateAppItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateAppItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "appitem.id",
            "description": "Unique ID for the appitem",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "appitem": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "title": "Name if items have names"
                    },
                    "description": {
                      "type": "string",
                      "title": "Description if appitem has a description"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "title": "Some tags"
                    },
                    "imageUrl": {
                      "type": "string",
                      "title": "A possible image url"
                    },
                    "difficulty": {
                      "type": "string",
                      "title": "Difficulty - exmaple attribute"
                    }
                  },
                  "title": "*\nAppItem being updated"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this AppItem to make partial changes."
                }
              },
              "title": "UpdateAppItemRequest"
            }
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      }
    },
    "/v1/appitems/{id}": {
      "get": {
        "summary": "GetAppItem returns a specific appitem with metadata",
        "operationId": "AppItemsService_GetAppItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAppItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular appitem",
        "operationId": "AppItemsService_DeleteAppItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteAppItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the appitem to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      }
    },
    "/v1/appitems:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple appitems by ID",
        "operationId": "AppItemsService_GetAppItems",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAppItemsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the appitem to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "AppItemsService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AppItem": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the appitem"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if appitem has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - exmaple attribute"
        }
      }
    },
    "v1CreateAppItemRequest": {
      "type": "object",
      "properties": {
        "appitem": {
          "$ref": "#/definitions/v1AppItem",
          "title": "*\nAppItem being updated"
        }
      },
      "title": "*\nAppItem creation request object"
    },
    "v1CreateAppItemResponse": {
      "type": "object",
      "properties": {
        "appitem": {
          "$ref": "#/definitions/v1AppItem",
          "title": "*\nAppItem being created"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an appitem creation."
    },
    "v1DeleteAppItemResponse": {
      "type": "object",
      "title": "*\nAppItem deletion response"
    },
    "v1GetAppItemResponse": {
      "type": "object",
      "properties": {
        "appitem": {
          "$ref": "#/definitions/v1AppItem"
        }
      }
    },
    "v1GetAppItemsResponse": {
      "type": "object",
      "properties": {
        "appitems": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1AppItem"
          }
        }
      },
      "title": "*\nAppItem batch-get response"
    },
    "v1ListAppItemsResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AppItem"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
    "v1Pagination": {
      "type": "object",
      "properties": {
        "pageKey": {
          "type": "string",
          "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set."
        },
        "pageOffset": {
          "type": "integer",
          "format": "int32",
          "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense."
        },
        "pageSize": {
          "type": "integer",
          "format": "int32",
          "description": "*\nNumber of results to return."
        }
      }
    },
    "v1PaginationResponse": {
      "type": "object",
      "properties": {
        "nextPageKey": {
          "type": "string",
          "description": "*\nThe key/pointer string that subsequent List requests should pass to\ncontinue the pagination."
        },
        "nextPageOffset": {
          "type": "integer",
          "format": "int32",
          "title": "*\nAlso support an integer offset if possible"
        },
        "hasMore": {
          "type": "boolean",
          "description": "*\nWhether theere are more results."
        },
        "totalResults": {
          "type": "integer",
          "format": "int32",
          "description": "*\nTotal number of results."
        }
      }
    },
    "v1UpdateAppItemResponse": {
      "type": "object",
      "properties": {
        "appitem": {
          "$ref": "#/definitions/v1AppItem",
          "title": "*\nAppItem being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an AppItem.",
      "title": "UpdateAppItemResponse"
    }
  }
}